refactor: flatten nested logic and split batch routing test for CodeScene
- Replace nested loops/if chains in spreadsheet_safety null-byte finders with
  small, single-purpose helpers and early-return dispatch.
- Simplify test_export_sink_guard AST and scan helpers using generator
  expressions and predicate helpers.
- Move shared batch-correction fixtures to conftest.py and split the routing
  test into test_batch_correction_routing.py to reduce file cohesion.
- All tests and flake8 pass.
