+++ scripts/tests/test_batch_correction.py
@@ -136,90 +136,6 @@ def patch_load_config(monkeypatch):
     yield
 
 
-@pytest.fixture
-def mock_dependencies(mocker):
-    """Mocks optional dependencies and file system calls."""
-    # Mock optional imports (assume they are NOT found by default)
-    mocker.patch("scripts.batch_correction.data_loader", None)
-    mocker.patch("scripts.batch_correction.processor", None)
-
-    # Mock file system interactions
-    mock_isdir = mocker.patch("os.path.isdir", return_value=True)
-    mock_isfile = mocker.patch("os.path.isfile", return_value=True)
-    mock_listdir = mocker.patch("os.listdir", return_value=[])
-
-    mock_getsize = mocker.patch("os.path.getsize", return_value=100)
-    mock_basename = mocker.patch(
-        "os.path.basename", side_effect=lambda p: os.path.split(p)[1]
-    )
-
-    # Mock pandas saving at the class level so the safe wrappers still route here
-    mock_to_excel = mocker.patch("pandas.DataFrame.to_excel")
-    mock_to_csv = mocker.patch("pandas.DataFrame.to_csv")
-
-    # Capture the real wrapper, then replace it with a recording mock that
-    # delegates to the real implementation so sanitization is still exercised.
-    from scripts import spreadsheet_safety as _ss
-
-    real_write_excel_safely = _ss.write_excel_safely
-    mock_write_excel_safely = mocker.patch(
-        "scripts.spreadsheet_safety.write_excel_safely"
-    )
-    mock_write_excel_safely.side_effect = real_write_excel_safely
-
-    # Mock builtins.open (for specific data file reads only)
-    mock_file_open = mocker.patch(
-        "builtins.open", mock.mock_open(read_data="line1\nline2")
-    )
-    mock_file_open.side_effect = None  # Reset side effect
-
-    return {
-        "isdir": mock_isdir,
-        "isfile": mock_isfile,
-        "listdir": mock_listdir,
-        "getsize": mock_getsize,
-        "basename": mock_basename,
-        "to_excel": mock_to_excel,
-        "to_csv": mock_to_csv,
-        "write_excel_safely": mock_write_excel_safely,
-        "open": mock_file_open,
-        "data_loader": None,  # Explicitly track mocked modules
-        "processor": None,
-    }
-
-
-@pytest.fixture
-def mock_config_loader(mocker):
-    """Provides a mock config loader function."""
-    mock_loader = mock.MagicMock(
-        return_value={
-            "RAW_DATA_DIR": "/fake/data/dir",
-            "RIVER_MILE_TO_SERIES": {"54.0": 26, "53.0": 27, "50.5": 28},
-            "RIVER_MILE_MAP_PATH": "scripts/river_mile_map.csv",
-        }
-    )
-    mocker.patch("scripts.batch_correction.load_config_func", mock_loader)
-    return mock_loader
-
-
-@pytest.fixture
-def mock_data_loader_mod(mocker):
-    """Provides a mock data_loader module."""
-    mock_mod = mock.MagicMock()
-    mock_mod.load_data.return_value = pd.DataFrame({0: range(5), 1: range(5)})
-    mocker.patch("scripts.batch_correction.data_loader", mock_mod)
-    return mock_mod
-
-
-@pytest.fixture
-def mock_processor_mod(mocker):
-    """Provides a mock processor module."""
-    mock_mod = mock.MagicMock()
-    mock_mod.process_data.return_value = pd.DataFrame({0: range(5), 1: range(5)})
-    mocker.patch("scripts.batch_correction.processor", mock_mod)
-    return mock_mod
-
-
 # --- Test Cases ---
 
 
@@ -722,47 +638,3 @@ def test_load_raw_data_empty_file(caplog):
         assert isinstance(result, pd.DataFrame)
         assert result.empty
         assert "dummy_empty_file.txt empty." in caplog.text
-
-
-def test_batch_process_routes_through_write_excel_safely(
-    mock_dependencies, mock_config_loader, mocker
-):
-    """The batch processor must route Excel exports through write_excel_safely
-    and the sanitizer must escape formula-like payloads before to_excel is called.
-    """
-    payload = '=HYPERLINK("http://attacker.example/collect","click")'
-
-    def read_csv_side_effect(path, *args, **kwargs):
-        if str(path).endswith("river_mile_map.csv"):
-            return pd.DataFrame({"SENSOR_ID": [26], "RIVER_MILE": [54.0]})
-        return pd.DataFrame(
-            {
-                0: [1, 2, 3],
-                1: [10.0, 20.0, 30.0],
-                2: [payload, "safe", "safe"],
-            }
-        )
-
-    mocker.patch("pandas.read_csv", side_effect=read_csv_side_effect)
-    mock_dependencies["listdir"].return_value = ["S26_Y01.txt"]
-    mock_dependencies["isfile"].return_value = True
-
-    summary_df = batch_process(
-        BatchConfig(
-            series_selection=26,
-            river_miles=None,
-            years=(1995, 1995),
-            dry_run=False,
-        )
-    )
-
-    assert len(summary_df) == 1
-    mock_write_excel = mock_dependencies["write_excel_safely"]
-    assert mock_write_excel.called
-
-    # The wrapper received the original DataFrame and the expected output path.
-    written_path = mock_write_excel.call_args.args[1]
-    assert written_path.endswith("Year_1995 (Y01)_Data.xlsx")
-    assert mock_write_excel.call_args.kwargs == {"index": False, "header": False}
-    df_passed = mock_write_excel.call_args.args[0]
-    assert df_passed.iloc[0, 2] == payload
